both cards extendes one class
